Extract testUtils
