+++ plugins/typescript/src/core/getDataResponseType.test.ts
@@ -1,5 +1,5 @@
 import { ResponsesObject } from "openapi3-ts";
-import ts from "typescript";
+import { print } from "../testUtils";
 import { getDataResponseType } from "./getDataResponseType";
 
 describe("getDataResponseType", () => {
@@ -127,14 +127,3 @@ describe("getDataResponseType", () => {
     expect(print(responseType)).toMatchInlineSnapshot(`"undefined"`);
   });
 });
-
-// Helpers
-const sourceFile = ts.createSourceFile("index.ts", "", ts.ScriptTarget.Latest);
-
-const printer = ts.createPrinter({
-  newLine: ts.NewLineKind.LineFeed,
-  removeComments: false,
-});
-
-const print = (node: ts.Node) =>
-  printer.printNode(ts.EmitHint.Unspecified, node, sourceFile);
