+++ plugins/typescript/src/testUtils.ts
@@ -1,0 +1,14 @@
+import ts from "typescript";
+
+const sourceFile = ts.createSourceFile("index.ts", "", ts.ScriptTarget.Latest);
+
+const printer = ts.createPrinter({
+  newLine: ts.NewLineKind.LineFeed,
+  removeComments: false,
+});
+
+/**
+ * Print a typescript node
+ */
+export const print = (node: ts.Node) =>
+  printer.printNode(ts.EmitHint.Unspecified, node, sourceFile);
